pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "spectra"
version = "0.1.0"
description = "Accessible OpenAPI documentation browser and REST client"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Spectra Contributors" }
]
dependencies = [
  "wxPython>=4.2",
  "requests>=2.32",
  "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-mock>=3.12",
  "ruff>=0.5"
]

[project.scripts]
spectra = "spectra.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/spectra"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
